+++ index.html
@@ -230,8 +230,6 @@
 
   <script src="/vendor/nouislider/nouislider.js"></script>
   <script src="/vendor/pristine/pristine.min.js"></script>
-  <script type="module" src="js/rendering-thumbnails.js"></script>
-  <script type="module" src="js/filters.js"></script>
   <script type="module" src="js/form.js"></script>
 </body>
 </html>
