+++ js/api.js
@@ -1,36 +1,48 @@
 const SERVER_URL = 'https://29.javascript.htmlacademy.pro/kekstagram';
 
-// получаем данные
+// Получение данных
 export const fetchPhotos = async () => {
   try {
     const response = await fetch(`${SERVER_URL}/data`);
     if (!response.ok) {
-      throw new Error('Не удалось загрузить данные');
+      throw new Error(`Ошибка: ${response.status} - ${response.statusText}`);
     }
     return await response.json();
   } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('Ошибка получения данных:', error);
+    document.body.insertAdjacentHTML('beforeend', `
+      <section class="error">
+        <div class="error__inner">
+          <h2 class="error__title">${error.message}</h2>
+          <button type="button" class="error__button" onclick="location.reload()">Попробовать ещё раз</button>
+        </div>
+      </section>
+    `);
     throw error;
   }
 };
 
-// отправляем данные
+// Отправка данных
 export const sendPhotoData = async (formData) => {
   try {
-    const response = await fetch(`${SERVER_URL}`, {
+    const response = await fetch(SERVER_URL, {
       method: 'POST',
       body: formData,
     });
 
     if (!response.ok) {
-      throw new Error('Не удалось отправить данные');
+      throw new Error(`Ошибка: ${response.status} - ${response.statusText}`);
     }
 
     return await response.json();
   } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('Ошибка отправки данных:', error);
+    document.body.insertAdjacentHTML('beforeend', `
+      <section class="error">
+        <div class="error__inner">
+          <h2 class="error__title">${error.message}</h2>
+          <button type="button" class="error__button">Попробовать ещё раз</button>
+        </div>
+      </section>
+    `);
     throw error;
   }
 };
